realtime-pipeline/cloud_functions/anomaly_detect/requirements.txt
google-cloud-pubsub==2.21.1
google-cloud-bigquery==3.17.2
functions-framework==3.5.0
